requirements.txt
jieba==0.42.1
numpy==2.3.5
pdfplumber==0.11.8
Pillow==12.0.0
pyside6==6.10.1
pyside6_addons==6.10.1
pyside6_essentials==6.10.1
python_docx==1.2.0
wordcloud==1.9.4
